requirements.txt
asgiref
bleach
certifi
cffi
chardet
cryptography
defusedxml
Django
django-allauth
django-crispy-forms
idna
importlib-metadata
Markdown
oauthlib
packaging
psycopg2
pycparser
PyJWT
pymdown-extensions
pyparsing
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
typing-extensions
urllib3
webencodings
zipp